+++ src/edit/administrativeNoteFields.js
@@ -1,20 +1,39 @@
+import { FieldArray } from 'react-final-form-arrays';
+import { Field } from 'react-final-form';
 import { FormattedMessage } from 'react-intl';
 
-import { TextArea } from '@folio/stripes/components';
+import {
+  RepeatableField,
+  Label,
+  TextArea,
+} from '@folio/stripes/components';
 
-import RepeatableField from '../components/RepeatableField';
+const administrativeNoteFields = () => {
+  const legend = (
+    <Label tagName="legend">
+      <FormattedMessage id="ui-inventory.administrativeNote" />
+    </Label>
+  );
 
-const administrativeNoteFields = () => (
-  <RepeatableField
-    name="administrativeNotes"
-    addLabel={<FormattedMessage id="ui-inventory.addAdministrativeNote" />}
-    addButtonId="clickable-add-administrative-note"
-    template={[{
-      component: TextArea,
-      label: <FormattedMessage id="ui-inventory.administrativeNote" />,
-      rows: 1
-    }]}
-  />
-);
+  const renderField = field => (
+    <Field
+      name={field}
+      component={TextArea}
+      rows={1}
+      fullWidth
+    />
+  );
+
+  return (
+    <FieldArray
+      name="administrativeNotes"
+      component={RepeatableField}
+      addLabel={<FormattedMessage id="ui-inventory.addAdministrativeNote" />}
+      onAdd={fields => fields.push('')}
+      headLabels={legend}
+      renderField={renderField}
+    />
+  );
+};
 
 export default administrativeNoteFields;