+++ src/edit/contributorFields.js
@@ -1,15 +1,26 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { FormattedMessage } from 'react-intl';
+import { FieldArray } from 'react-final-form-arrays';
+import { Field } from 'react-final-form';
 import {
+  FormattedMessage,
+  useIntl,
+} from 'react-intl';
+
+import {
+  RepeatableField,
   TextArea,
   Select,
+  Row,
+  Col,
+  Label,
 } from '@folio/stripes/components';
-import { IntlConsumer } from '@folio/stripes/core';
-import RepeatableField from '../components/RepeatableField';
+
 import PrimaryToggleButton from './components/PrimaryToggleButton';
 
 const ContributorFields = props => {
+  const { formatMessage } = useIntl();
+
   const {
     contributorNameTypes,
     contributorTypes,
@@ -27,61 +38,102 @@ const ContributorFields = props => {
     value: it.id,
   }));
 
-  return (
-    <IntlConsumer>
-      {intl => (
-        <RepeatableField
-          name="contributors"
-          label={<FormattedMessage id="ui-inventory.contributors" />}
-          addLabel={<FormattedMessage id="ui-inventory.addContributor" />}
-          addButtonId="clickable-add-contributor"
-          template={[
-            {
-              label: <FormattedMessage id="ui-inventory.name" />,
-              name: 'name',
-              component: TextArea,
-              rows: 1,
-              required: true,
-              disabled: !canEdit,
-            }, {
-              label: <FormattedMessage id="ui-inventory.nameType" />,
-              name: 'contributorNameTypeId',
-              component: Select,
-              placeholder: intl.formatMessage({ id: 'ui-inventory.selectType' }),
-              dataOptions: contributorNameTypeOptions,
-              required: true,
-              disabled: !canEdit,
-            }, {
-              label: <FormattedMessage id="ui-inventory.type" />,
-              name: 'contributorTypeId',
-              component: Select,
-              placeholder: intl.formatMessage({ id: 'ui-inventory.selectType' }),
-              dataOptions: contributorTypeOptions,
-              disabled: !canEdit,
-            }, {
-              label: <FormattedMessage id="ui-inventory.typeFreeText" />,
-              name: 'contributorTypeText',
-              rows: 1,
-              component: TextArea,
-              disabled: !canEdit,
-            }, {
-              name: 'primary',
-              label: intl.formatMessage({ id: 'ui-inventory.primary' }),
-              component: PrimaryToggleButton,
-              disabled: !canEdit,
-            },
-          ]}
-          newItemTemplate={{
-            name: '',
-            contributorNameTypeId: '',
-            primary: false,
-            contributorTypeText: '',
-          }}
-          canAdd={canAdd}
-          canDelete={canDelete}
+  const headLabels = (
+    <Row>
+      <Col sm={2}>
+        <Label tagName="legend" required>
+          <FormattedMessage id="ui-inventory.name" />
+        </Label>
+      </Col>
+      <Col sm={2}>
+        <Label tagName="legend" required>
+          <FormattedMessage id="ui-inventory.nameType" />
+        </Label>
+      </Col>
+      <Col sm={2}>
+        <Label tagName="legend">
+          <FormattedMessage id="ui-inventory.type" />
+        </Label>
+      </Col>
+      <Col sm={4}>
+        <Label tagName="legend">
+          <FormattedMessage id="ui-inventory.typeFreeText" />
+        </Label>
+      </Col>
+      <Col sm={2}>
+        <Label tagName="legend">
+          <FormattedMessage id="ui-inventory.primary" />
+        </Label>
+      </Col>
+    </Row>
+  );
+
+  const renderField = (field, index, fields) => (
+    <Row>
+      <Col sm={2}>
+        <Field
+          name={`${field}.name`}
+          component={TextArea}
+          rows={1}
+          disabled={!canEdit}
+          required
         />
-      )}
-    </IntlConsumer>
+      </Col>
+      <Col sm={2}>
+        <Field
+          name={`${field}.contributorNameTypeId`}
+          component={Select}
+          placeholder={formatMessage({ id: 'ui-inventory.selectType' })}
+          dataOptions={contributorNameTypeOptions}
+          disabled={!canEdit}
+          required
+        />
+      </Col>
+      <Col sm={2}>
+        <Field
+          name={`${field}.contributorTypeId`}
+          component={Select}
+          placeholder={formatMessage({ id: 'ui-inventory.selectType' })}
+          dataOptions={contributorTypeOptions}
+          disabled={!canEdit}
+        />
+      </Col>
+      <Col sm={4}>
+        <Field
+          name={`${field}.contributorTypeText`}
+          component={TextArea}
+          rows={1}
+          disabled={!canEdit}
+        />
+      </Col>
+      <Col sm={2}>
+        <Field
+          name={`${field}.primary`}
+          component={PrimaryToggleButton}
+          disabled={!canEdit}
+          fields={fields}
+        />
+      </Col>
+    </Row>
+  );
+
+  return (
+    <FieldArray
+      name="contributors"
+      component={RepeatableField}
+      legend={<FormattedMessage id="ui-inventory.contributors" />}
+      addLabel={<FormattedMessage id="ui-inventory.addContributor" />}
+      onAdd={fields => fields.push({
+        name: '',
+        contributorNameTypeId: '',
+        contributorTypeText: '',
+        primary: false,
+      })}
+      headLabels={headLabels}
+      renderField={renderField}
+      canAdd={canAdd}
+      canRemove={canDelete}
+    />
   );
 };