+++ src/edit/publicationFields.js
@@ -1,10 +1,17 @@
 import React from 'react';
 import { FormattedMessage } from 'react-intl';
+import { FieldArray } from 'react-final-form-arrays';
+import { Field } from 'react-final-form';
 import PropTypes from 'prop-types';
 
-import { TextField, TextArea } from '@folio/stripes/components';
-
-import RepeatableField from '../components/RepeatableField';
+import {
+  RepeatableField,
+  TextField,
+  TextArea,
+  Row,
+  Col,
+  Label,
+} from '@folio/stripes/components';
 
 const PublicationFields = props => {
   const {
@@ -13,42 +20,83 @@ const PublicationFields = props => {
     canDelete,
   } = props;
 
+
+  const headLabels = (
+    <Row>
+      <Col sm={3}>
+        <Label tagName="legend">
+          <FormattedMessage id="ui-inventory.publisher" />
+        </Label>
+      </Col>
+      <Col sm={3}>
+        <Label tagName="legend">
+          <FormattedMessage id="ui-inventory.publisherRole" />
+        </Label>
+      </Col>
+      <Col sm={3}>
+        <Label tagName="legend">
+          <FormattedMessage id="ui-inventory.place" />
+        </Label>
+      </Col>
+      <Col sm={3}>
+        <Label tagName="legend">
+          <FormattedMessage id="ui-inventory.dateOfPublication" />
+        </Label>
+      </Col>
+    </Row>
+  );
+
+  const renderField = field => (
+    <Row>
+      <Col sm={3}>
+        <Field
+          name={`${field}.publisher`}
+          component={TextArea}
+          rows={1}
+          disabled={!canEdit}
+        />
+      </Col>
+      <Col sm={3}>
+        <Field
+          name={`${field}.role`}
+          component={TextField}
+          disabled={!canEdit}
+        />
+      </Col>
+      <Col sm={3}>
+        <Field
+          name={`${field}.place`}
+          component={TextArea}
+          rows={1}
+          disabled={!canEdit}
+        />
+      </Col>
+      <Col sm={3}>
+        <Field
+          name={`${field}.dateOfPublication`}
+          component={TextField}
+          disabled={!canEdit}
+        />
+      </Col>
+    </Row>
+  );
+
   return (
-    <RepeatableField
+    <FieldArray
       name="publication"
-      label={<FormattedMessage id="ui-inventory.publications" />}
+      component={RepeatableField}
+      legend={<FormattedMessage id="ui-inventory.publications" />}
       addLabel={<FormattedMessage id="ui-inventory.addPublication" />}
-      addButtonId="clickable-add-publication"
-      template={[
-        {
-          name: 'publisher',
-          label: <FormattedMessage id="ui-inventory.publisher" />,
-          component: TextArea,
-          rows: 1,
-          disabled: !canEdit,
-        },
-        {
-          name: 'role',
-          label: <FormattedMessage id="ui-inventory.publisherRole" />,
-          component: TextField,
-          disabled: !canEdit,
-        },
-        {
-          name: 'place',
-          label: <FormattedMessage id="ui-inventory.place" />,
-          component: TextArea,
-          rows: 1,
-          disabled: !canEdit,
-        },
-        {
-          name: 'dateOfPublication',
-          label: <FormattedMessage id="ui-inventory.dateOfPublication" />,
-          component: TextField,
-          disabled: !canEdit,
-        },
-      ]}
+      onAdd={fields => fields.push({
+        publisher: '',
+        role: '',
+        place: '',
+        dateOfPublication: '',
+      })}
+      headLabels={headLabels}
+      renderField={renderField}
       canAdd={canAdd}
-      canDelete={canDelete}
+      canRemove={canDelete}
     />
   );
 };