+++ src/edit/seriesFields.js
@@ -1,10 +1,14 @@
 import React from 'react';
+import { FieldArray } from 'react-final-form-arrays';
+import { Field } from 'react-final-form';
 import { FormattedMessage } from 'react-intl';
 import PropTypes from 'prop-types';
 
-import { TextArea } from '@folio/stripes/components';
-
-import RepeatableField from '../components/RepeatableField';
+import {
+  Label,
+  RepeatableField,
+  TextArea,
+} from '@folio/stripes/components';
 
 const SeriesFields = props => {
   const {
@@ -13,21 +17,31 @@ const SeriesFields = props => {
     canDelete,
   } = props;
 
+  const legend = (
+    <Label tagName="legend">
+      <FormattedMessage id="ui-inventory.seriesStatements" />
+    </Label>
+  );
+
+  const renderField = field => (
+    <Field
+      name={`${field}.value`}
+      component={TextArea}
+      rows={1}
+      disabled={!canEdit}
+    />
+  );
+
   return (
-    <RepeatableField
+    <FieldArray
       name="series"
-      label={<FormattedMessage id="ui-inventory.seriesStatements" />}
+      component={RepeatableField}
       addLabel={<FormattedMessage id="ui-inventory.addSeries" />}
-      addButtonId="clickable-add-series"
-      template={[{
-        component: TextArea,
-        disabled: !canEdit,
-        rows: 1,
-        name: 'value'
-      }]}
+      onAdd={fields => fields.push({ value: '' })}
+      headLabels={legend}
+      renderField={renderField}
       canAdd={canAdd}
-      canDelete={canDelete}
-      newItemTemplate={{ value: '' }}
+      canRemove={canDelete}
     />
   );
 };